Move config cache tests into own test package and some cleanup
