+++ src/main/resources/org/gradlex/javamodule/dependencies/unique_modules.properties
@@ -1151,7 +1151,7 @@ com.github.gv2011.util.gcol=com.github.gv2011:util-gcol
 com.github.gv2011.util.html.imp=com.github.gv2011:util-html
 com.github.gv2011.util.image=com.github.gv2011:util-image
 com.github.gv2011.util.json.imp=com.github.gv2011:util-json
-com.github.gv2011.util.swing=com.github.gv2011:util-swing
+com.github.gv2011.util.swing.imp=com.github.gv2011:util-swing
 com.github.gv2011.util.tika=com.github.gv2011:util-tika
 com.github.gw2toolbelt.gw2ml=com.github.gw2toolbelt.gw2ml:gw2ml
 com.github.hanshsieh.pixivj=com.github.hanshsieh:pixivj
