+++ src/test/groovy/org/gradlex/javamodule/dependencies/test/initialization/SettingsPluginTest.groovy
@@ -1,6 +1,5 @@
 package org.gradlex.javamodule.dependencies.test.initialization
 
-import org.gradle.testkit.runner.GradleRunner
 import org.gradlex.javamodule.dependencies.test.fixture.GradleBuild
 import spock.lang.Specification
 
@@ -91,37 +90,7 @@ class SettingsPluginTest extends Specification {
         result.getOutput().contains("Calculating task graph as no cached configuration is available for tasks: :app:compileJava")
 
         when:
-        result = runner.build() //TODO Thats a big problem with the Settings Plugin AAAAAAAAH
-        result = runner.build()
-
-        then:
-        result.getOutput().contains("Reusing configuration cache.")
-    }
-
-    def "configurationCacheHit"() {
-        given:
-        settingsFile << '''
-            javaModules {
-                directory(".") { plugin("java-library") }
-            }
-        '''
-        libModuleInfoFile << 'module abc.lib { }'
-        appModuleInfoFile << '''
-            module org.gradlex.test.app {
-                requires abc.lib;
-            }
-        '''
-
-
-        def runner = runner(':app:compileJava')
-        when:
-        def result = runner.build()
-
-        then:
-        result.getOutput().contains("Calculating task graph as no cached configuration is available for tasks: :app:compileJava")
-
-        when:
-        result = runner.build() //TODO Thats a big problem with the Settings Plugin AAAAAAAAH
+        runner.build() // https://github.com/gradlex-org/java-module-dependencies/issues/128
         result = runner.build()
 
         then:
@@ -150,7 +119,7 @@ class SettingsPluginTest extends Specification {
         result.getOutput().contains("Calculating task graph as no cached configuration is available for tasks: :app:compileJava")
 
         when:
-        result = runner.build() //TODO Thats a big problem with the Settings Plugin AAAAAAAAH
+        runner.build() // https://github.com/gradlex-org/java-module-dependencies/issues/128
         appModuleInfoFile.write('''
             module org.gradlex.test.app {
                 requires abc.lib; //This is a comment and should not break the configurationCache
@@ -184,7 +153,7 @@ class SettingsPluginTest extends Specification {
         result.getOutput().contains("Calculating task graph as no cached configuration is available for tasks: :app:compileJava")
 
         when:
-        result = runner.build() //TODO Thats a big problem with the Settings Plugin AAAAAAAAH
+        runner.build() // https://github.com/gradlex-org/java-module-dependencies/issues/128
         appModuleInfoFile.write('''
             module org.gradlex.test.app {
                //dependency removed; so thats indeed a configuration change
